+++ src/scss/main.scss
@@ -1,2 +1,2 @@
-@import './basket.scss';
-@import './other.scss';
+@import './normalize.scss';
+@import'./style.scss';
