--- src/scss/other.scss
@@ -1,54 +1,0 @@
-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background: rgb(32, 32, 32);
-}
-
-h1 {
-    color: #fff;;
-}
-
-.btn-cart {
-    height: 10px;
-}
-
-.products {
-    margin-top: 20px;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-}
-
-.product-item {
-    display: flex;
-    /* border: 1px solid; */
-    margin-bottom: 10px;
-    flex-direction: column;
-    justify-content: space-between;
-    font-size: 20px;
-    flex-grow: 0;
-}
-
-.item-img{
-    flex-grow: 1;
-}
-
-.item-img img{
-    width: 200px;
-}
-
-h3 {
-    font-size: 38px;
-}
-
-p, .buy-btn {
-    font-size: 25px;
-}
-
-.btn-active:active {
-    box-shadow:
-    inset 0 1px 3px rgba(0,10,20,.5),
-    0 1px rgb(83,94,104),
-    0 0 1px rgb(86,96,106);
-}
