--- src/xhr.js
@@ -1,18 +1,0 @@
-function makeGETRequest(url, callback) {
-    var xhr;
-  
-    if (window.XMLHttpRequest) {
-      xhr = new XMLHttpRequest();
-    } else if (window.ActiveXObject) { 
-      xhr = new ActiveXObject("Microsoft.XMLHTTP");
-    }
-  
-    xhr.onreadystatechange = function () {
-      if (xhr.readyState === 4) {
-        callback(xhr.responseText);
-      }
-    }
-  
-    xhr.open('GET', url, true);
-    xhr.send();
-  }
